feat: move combinations to new file with better docs
