+++ src/js/base/index.js
@@ -5,6 +5,7 @@ const build = require("./build");
 const compat = require("./compat");
 const errors = require("./errors");
 const interactable = require("./interactable");
+const itertools = require("./itertools");
 const logic = require("./logic");
 const config = require("./config");
 const mobile = require("./mobile");
@@ -20,6 +21,7 @@ Object.assign(module.exports, build);
 Object.assign(module.exports, compat);
 Object.assign(module.exports, errors);
 Object.assign(module.exports, interactable);
+Object.assign(module.exports, itertools);
 Object.assign(module.exports, logic);
 Object.assign(module.exports, config);
 Object.assign(module.exports, mobile);
